Moved byte and bmp tests to the separate module
