+++ build.sbt
@@ -132,6 +132,14 @@ lazy val cssparse = crossProject
 lazy val cssparseJVM = cssparse.jvm
 lazy val cssparseJS = cssparse.js
 
+lazy val byteparse = crossProject.dependsOn(fastparse).settings(
+  name := "byteparse"
+).settings(shared:_*)
+  .jvmSettings()
+
+lazy val byteparseJVM = byteparse.jvm
+lazy val byteparseJS = byteparse.js
+
 lazy val perftests = crossProject.settings(
   name := "perfomance-tests"
 ).settings(shared:_*).dependsOn(fastparse, pythonparse, scalaparse, cssparse)
@@ -147,6 +155,8 @@ lazy val modules = project.aggregate(
   cssparseJVM,
   scalaparseJS,
   scalaparseJVM,
+  byteparseJVM,
+  byteparseJS,
   utilsJS,
   utilsJVM
 ).settings(
