+++ byteparse/jvm/src/test/resources/inverted_lena.bmp
