+++ byteparse/jvm/src/test/resources/lena.bmp
